Fix terminating stopped RDS instances (#319)
RDS instances that have deletion protection enabled must be modified to
disable this before they can be deleted. If the instance is stopped, it
has to be restarted before it can be modifed, though. This can take quite
a few minutes, so it's not feasible to wait for an instance to restart in
a single terminator run. A restart will have to be issued and then the
deletion picked back up on a subsequent terminator run.